+++ src/lib.rs
@@ -67,7 +67,7 @@ pub use blob::Entry;
 pub use error::{Error, Kind as ErrorKind};
 pub use record::Meta;
 pub use rio;
-pub use storage::{ArrayKey, Builder, Key, RefKey, Storage};
+pub use storage::{ArrayKey, BlobRecordTimestamp, Builder, Key, ReadResult, RefKey, Storage};
 
 mod prelude {
     use crc::{Crc, CRC_32_ISCSI};
@@ -81,11 +81,7 @@ mod prelude {
     pub(crate) use bincode::{deserialize, serialize, serialize_into, serialized_size};
     pub(crate) use blob::{self, Blob, IndexConfig};
     pub(crate) use filter::{Bloom, BloomProvider, Config as BloomConfig, HierarchicalFilters};
-    pub(crate) use futures::{
-        future,
-        lock::Mutex,
-        stream::{futures_unordered::FuturesUnordered, TryStreamExt},
-    };
+    pub(crate) use futures::{lock::Mutex, stream::futures_unordered::FuturesUnordered};
     pub(crate) use record::{Header as RecordHeader, Record, RECORD_MAGIC_BYTE};
     pub(crate) use rio::Rio;
     pub(crate) use std::{